ASL_Project/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow-macos
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
scikit-learn==1.3.0
matplotlib==3.7.2
tensorflow-metal